+++ spec/dummy/app/channels/widget_channel.rb
@@ -7,6 +7,13 @@ class WidgetChannel < ActionCable::Channel::Base
     transmit({ status: "following", id: data["id"] })
   end
 
+  def observed(data)
+    Widget.where(id: data["id"]).count
+    Rails.logger.info("observed widget channel action")
+    ActionCable.server.broadcast("widgets:#{data['id']}:updates", { status: "observed" })
+    transmit({ status: "observed", id: data["id"] })
+  end
+
   def explode(_data)
     raise "channel action failed"
   end